Circuit sketch for vocal_numeric
IMPORTANT NOTE: This circuit uses two different voltages. The Arduino Nano 33 BLE Sense uses 5V power while the shift register and display are using 3.3V to match the Nano's logic output. This is to maintain the ability of the development board to use USB power and input. In production, the entire circuit would use 3.3V